feat: added io/roundtrip.go to factorize code
